bulk-sms-system/bulk_sms/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
dj-database-url==2.3.0
Django==4.2.20
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==9.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
whitenoise==6.9.0
africastalking==1.2.7
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
twilio==8.5.0

# Template Verification Dependencies
profanity-check==1.0.3
scikit-learn>=1.6.1
numpy>=2.2.4
scipy>=1.15.2
joblib>=1.4.2
threadpoolctl>=3.6.0
